+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+### Fixed
+- Update textile-js to GehDoc/textile-js@5a32824bdc6e806615a0d45003f0d32e5c2bba48. Fixes #30 and #31.
+
+
 ## [0.3.6] - 20200201
 ### Added
 - [SOURCE] Implement links and image extraction and activation